Main

See merge request groep-t/courses/apt/2324/team-a4-fa!29

+++ Model/ProtagonistModel.cpp
@@ -51,7 +51,7 @@ void ProtagonistModel::setHealth(float health) {
     protagonist->setHealth(health);
 }
 
-int ProtagonistModel::getEnergy() const {
+float ProtagonistModel::getEnergy() const {
     return protagonist->getEnergy();
 }
